fixup! fix(ivy): align NgModule registration timing with ViewEngine
